Guard VC settings retrieval against environment variables.
In some cases %PATH% is cleaned up by a platform SDK (the PS4 SDK
appears to be causing that, though I am not 100% sure). In this case
the VCEnvironment.SetEnvironment() call will throw. However we can
ignore the failure since it most probably means we are not going to
build with the Visual Studio compiler.